Now Level provides a bounds-enforced indexed lens.
This has been done because in general having out-of-bounds
coordinates will be the sign of a bug, so when this
happens a lot of noise should be made.  Furthermore,
given that we are assuming in-bounds coordinates, we can
make the indexed lens be a Lens rather than merely a
Traversable.